requirements.txt
-i https://pypi.org/simple
argcomplete==3.1.4
blinker==1.7.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
commitizen==3.12.0
decli==0.6.1
flask==3.0.0
flask-wtf==1.2.1
importlib-metadata==6.8.0
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3
packaging==23.2
prompt-toolkit==3.0.36
python-dotenv==1.0.0
pyyaml==6.0.1
questionary==2.0.1
termcolor==2.3.0
tomlkit==0.12.2
wcwidth==0.2.9
werkzeug==3.0.1
wtforms==3.1.1
zipp==3.17.0
